test(cli): table-ize command-suggestion units, split out built-CLI integration tests
Addresses review: extractCommandTokens + suggestCommand cases become it.each
tables; the built-CLI/dummy-server tests (real command_not_found hook +
telemetry transport) move to cli-command-not-found.integration.test.ts,
leaving the pure-function unit tests in cli-command-not-found.test.ts.
